Don't use scratchpad to pass values into partials We previously used .Scratch which is a global scratchpad, that was later passed over to the partial with the rest of the context. The partial then checked if this value was set, and had a fallback if it wasn't (default value).
This approach does work in most of the cases, however in about 5% of
runs, it fails. I assume this is due to hugo spinning up multiple
threads, and each of them is changing this global scratchpad in
arbitrary order. At some point, a thread then removed this value from
the scratchpad to reset it, but another thread already checked that this
variable does exist, and has now set the no longer available variable.

To avoid this, we can instead directly pass a custom dict, which
contains the original context (.) along with another variable, which
will hold the dateformat. This however means that we can't use the
simple syntax with a default value, and the dateformat will need to be
set each time (though technically, it could be set to nil and checked
later, but if we're already specifying it, we may as well pass in this
dateformat directly)